requirements.txt
claude-code-sdk
anthropic
faker
pydantic
python-dotenv
pytest
